type: feature.
Retrieved token from user table from DB.
Added getProspects logic with Rx.
Response stored on a list (for future recyclerView)
Response added on DB too with Rx.

Note: As the JSON from the original rest service was amorphous; retrofit
and other tested libraries where presenting an error a where unable to
store the response on the list<>. For this exercise i used a mock rest
service that i create with mockable.io and fixing the JSON Manually.